+++ src/animation/use-animate-prop.ts
@@ -1,5 +1,5 @@
 import { useRef, useEffect } from "react"
-import { Target, Transition } from "../types"
+import { Target, Transition, TargetAndTransition } from "../types"
 import { ComponentAnimationControls } from "./ComponentAnimationControls"
 import { MotionValuesMap } from "../motion/utils/use-motion-values"
 import { shallowCompare } from "../utils/use-inline"
@@ -38,17 +38,18 @@ export const hasUpdated = (
  * @internal
  */
 export function useAnimateProp(
-    target: Target,
+    target: TargetAndTransition,
     controls: ComponentAnimationControls,
     values: MotionValuesMap,
-    transition?: Transition
+    defaultTransition?: Transition
 ) {
     const isInitialRender = useRef(true)
     const prevValues = useRef(target)
 
     useEffect(
         () => {
-            const toAnimate: Target = {}
+            const animateTarget: Target = {}
+            const { transition, transitionEnd } = target
 
             for (const key in prevValues.current) {
                 const shouldAnimateOnMount =
@@ -59,7 +60,7 @@ export function useAnimateProp(
                     hasUpdated(prevValues.current[key], target[key]) ||
                     shouldAnimateOnMount
                 ) {
-                    toAnimate[key] = target[key]
+                    animateTarget[key] = target[key]
                 }
             }
 
@@ -69,8 +70,12 @@ export function useAnimateProp(
                 ...target,
             }
 
-            if (Object.keys(toAnimate).length) {
-                controls.start(toAnimate, transition)
+            if (Object.keys(animateTarget).length) {
+                controls.start({
+                    ...animateTarget,
+                    transition: transition || defaultTransition,
+                    transitionEnd,
+                })
             }
         },
         [target]