+++ src/motion/__tests__/animate-prop.test.tsx
@@ -42,6 +42,67 @@ describe("animate prop as object", () => {
         return expect(promise).resolves.toBe(50)
     })
 
+    test("uses transition on subsequent renders", async () => {
+        const promise = new Promise(resolve => {
+            const x = motionValue(0)
+            const Component = ({ animate }: any) => (
+                <motion.div animate={animate} style={{ x }} />
+            )
+            const { rerender } = render(
+                <Component animate={{ x: 10, transition: { type: false } }} />
+            )
+            rerender(
+                <Component animate={{ x: 20, transition: { type: false } }} />
+            )
+            rerender(
+                <Component animate={{ x: 30, transition: { type: false } }} />
+            )
+
+            requestAnimationFrame(() => resolve(x.get()))
+        })
+
+        return expect(promise).resolves.toBe(30)
+    })
+
+    test("uses transitionEnd on subsequent renders", async () => {
+        const promise = new Promise(resolve => {
+            const x = motionValue(0)
+            const Component = ({ animate }: any) => (
+                <motion.div animate={animate} style={{ x }} />
+            )
+            const { rerender } = render(
+                <Component
+                    animate={{
+                        x: 10,
+                        transition: { type: false },
+                        transitionEnd: { x: 100 },
+                    }}
+                />
+            )
+            rerender(
+                <Component
+                    animate={{
+                        x: 20,
+                        transition: { type: false },
+                        transitionEnd: { x: 200 },
+                    }}
+                />
+            )
+            rerender(
+                <Component
+                    animate={{
+                        x: 30,
+                        transition: { type: false },
+                        transitionEnd: { x: 300 },
+                    }}
+                />
+            )
+            requestAnimationFrame(() => resolve(x.get()))
+        })
+
+        return expect(promise).resolves.toBe(300)
+    })
+
     test("animates to set prop and preserves existing initial transform props", async () => {
         const promise = new Promise(resolve => {
             const onComplete = () => {